refactor: test 파일 refactoring.
